+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Elixir standard library modules
 - ESP32: `--boot` option to mkimage.sh tool
 - Add `erlang:atom_to_binary/1` that is equivalent to `erlang:atom_to_binary(Atom, utf8)`
+- Support for Elixir `String.Chars` protocol
 
 ### Changed